+++ std/lib/treap_node.h
@@ -1,7 +1,5 @@
 #pragma once
 
-#include <std/sys/types.h>
-
 namespace Std {
     struct TreapVisitor;
 
@@ -12,6 +10,6 @@ namespace Std {
         virtual void* key() const noexcept;
 
         void visit(TreapVisitor& vis);
-        size_t height() const noexcept;
+        unsigned height() const noexcept;
     };
 }
